move flat_map_identity to its own module
